% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce2data.table.R
\name{coerce2data.table}
\alias{coerce2data.table}
\title{coerce2data.table}
\usage{
coerce2data.table(df, keep.rownames = FALSE, check.names = FALSE,
  key = NULL, stringsAsFactors = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{keep.rownames}{see \link[data.table]{data.table}}

\item{check.names}{see \link[data.table]{data.table}}

\item{key}{see \link[data.table]{data.table}}

\item{stringsAsFactors}{see \link[data.table]{data.table}}
}
\value{
data.table
}
\description{
Coerces \code{data.frame} to \code{data.table} class
}
