% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{compareImages}
\alias{compareImages}
\title{compareImages}
\usage{
compareImages(x, y, tolerance = 0L)
}
\arguments{
\item{x}{path to image}

\item{y}{path to image}

\item{tolerance}{allowed percentage of difference}
}
\value{
logical
}
\description{
Compare two images
}
