% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderBoxplot.R
\name{renderBoxplot}
\alias{renderBoxplot}
\title{renderBoxplot}
\usage{
renderBoxplot(dat, jitter = 0.2)
}
\arguments{
\item{dat}{data.frame}

\item{jitter}{numeric}
}
\value{
\code{plotly} object
}
\description{
Render boxplot
}
\examples{
\dontrun{
dat <- data.frame(xval = sample(100, 1000, replace = TRUE),
                  group = as.factor(sample(c("a", "b", "c"), 1000, replace = TRUE)))

renderBoxplot(dat)
}
}
