% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{takeScreenshot}
\alias{takeScreenshot}
\title{takeScreenshot}
\usage{
takeScreenshot(EXPR, target.path = tempfile(fileext = ".png"),
  overwrite = TRUE, is.plotly = inherits(EXPR, "plotly"))
}
\arguments{
\item{EXPR}{R expression to evaluate}

\item{target.path}{character; location of output file}

\item{overwrite}{logical (default \code{TRUE}); if \code{TRUE} then existing file on \code{target.path} will be overwritten}

\item{is.plotly}{logical; if \code{TRUE} then use function \link[plotly]{export} to save output;
tries to guess using \code{inherits(p, "plotly")} by default}
}
\value{
path to file
}
\description{
Takes screenshot from evaluated expressions' output.
}
\author{
Kamil Foltyński
}
